backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
alembic==1.13.2
pydantic==2.7.4
pydantic-settings==2.3.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
httpx==0.27.0
docker==7.1.0
stripe==10.8.0
